app/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi[standard]>=0.115.2",
    "openai",
    "openpyxl>=3.1.5",
    "pymupdf>=1.1.2",
    "python-docx",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.12.0",
    "structlog",
    "tiktoken>=0.8.0",
    "uvicorn",
]
